requirements.txt
fastapi
uvicorn
requests
pandas
psycopg2-binary
openai
python-dotenv
pytest